requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.5
scikit-learn>=1.3.2
faiss-gpu>=1.7.3  
infomap>=2.7.1
tqdm>=4.65.0
timm==0.4.12
Pillow>=9.4.0
